+++ package.json
@@ -6,7 +6,7 @@
     "cra-template": "1.2.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^7.1.3",
+    "react-router-dom": "^6.28.2",
     "react-scripts": "^3.0.1",
     "web-vitals": "^4.2.4"
   },
@@ -33,8 +33,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "webpack": "^4.42.0"
   }
 }
