+++ src/pages/postHS.js
@@ -1,0 +1,315 @@
+import React from "react";
+import "./postHS.css";
+
+// 이미지 import
+import logo from "../images/logo.svg";
+import arrowBottom from "../images/arrow_bottom.svg";
+import addEmoji from "../images/ico_add.svg";
+import shareIcon from "../images/share-24.svg";
+import plusIcon from "../images/plus.svg";
+import deleteIcon from "../images/ico_delete.svg";
+
+// 배지 컴포넌트
+const Badge = ({ type }) => {
+  const BADGE_STYLES = {
+    지인: { background: "#FFF0D6", color: "#FF8832" }, // 연한 주황색
+    동료: { background: "#F8F0FF", color: "#9935FF" }, // 연한 보라색
+    가족: { background: "#E4FBDC", color: "#2BA600" }, // 연한 초록색
+    친구: { background: "#E2F5FF", color: "#00A2FE" }, // 연한 파란색
+  };
+
+  return (
+    <em className="badge" style={BADGE_STYLES[type]}>
+      {type}
+    </em>
+  );
+};
+
+// 이모티콘과 그에 대한 반응 카운트
+const EMOJI_DATA = [
+  { emoji: "🥰", count: 24 },
+  { emoji: "😂", count: 16 },
+  { emoji: "😎", count: 10 },
+];
+
+// 각 문장마다 다른 폰트 적용하기
+const FONT_STYLES = {
+  notoSans: { fontFamily: '"Noto-Sans", sans-serif' },
+  pretendard: { fontFamily: '"Pretendard", sans-serif' },
+  nanumMyeongjo: { fontFamily: '"NanumMyengjo", serif' },
+  NanumSonPyeonJiCe: { fontFamily: '"NanumSonPyeonJiCe", sans-serif', fontSize: "24px" },
+};
+
+function Post() {
+
+  return (
+    <>
+      <header>
+        <div className="container">
+          <a href="#" className="logo">
+            <img src={logo} alt="롤링페이퍼 로고" />
+          </a>
+          <a href="#" className="btn_making">롤링페이퍼 만들기</a>
+        </div>
+      </header>
+      <main>
+        <h1 className="blind">보낸 롤링페이퍼 리스트</h1>
+        {/* postHeader 시작 */}
+        <div className="postHeader">
+          <div className="container">
+            <div className="leftWrap">
+              <p>To. Ashley Kim</p>
+            </div>
+            <div className="rightWrap">
+              <div className="emojiReactionWrap">
+                <div className="emojiCollection">
+                  <ul className="emojiTop3List">
+                    {EMOJI_DATA.map((emoji, index) => (
+                      <li key={index}>
+                        <em>{emoji.emoji}</em><span>{emoji.count}</span>
+                      </li>
+                    ))}
+                  </ul>
+                  <div className="emojiAllList">
+                    <button>
+                      <img src={arrowBottom} alt="이모지 전체보기" />
+                    </button>
+                    <ul>
+                      {EMOJI_DATA.map((emoji, index) => (
+                        <li key={index}>
+                          <em>{emoji.emoji}</em><span>{emoji.count}</span>
+                        </li>
+                      ))}
+                    </ul>
+                  </div>
+                </div>
+                <div className="emojiPicker">
+                  <button>
+                    <img src={addEmoji} alt="이모지 추가하기" /><span>추가</span>
+                  </button>
+                  <div></div>
+                </div>
+              </div>
+              <div className="shareSnsWrap">
+                <button>
+                  <img src={shareIcon} alt="공유하기" />
+                </button>
+                <ul>
+                  <li><a href="#">카카오톡 공유</a></li>
+                  <li><a href="#">URL 공유</a></li>
+                </ul>
+              </div>
+            </div>
+          </div>
+        </div>
+        {/* postHeader 끝 */}
+
+        <div className="post">
+          <div className="container">
+            <ul className="postCard">
+              <li className="addPostCard">
+                <a href="#">
+                  <span>
+                    <img src={plusIcon} alt="이모지 추가하기" />
+                  </span>
+                </a>
+              </li>
+              <li className="savedPostCard">
+                <a href="#">
+                  <div className="cardInfo">
+                    <div>
+                      <div className="photo"></div>
+                      <div className="fromName">
+                        <span>From. <em>김동훈</em></span>
+                        <Badge type="친구" />
+                      </div>
+                    </div>
+                    <a className="btnDelete">
+                      <img src={deleteIcon} alt="삭제하기" />
+                    </a>
+                  </div>
+                  <p className="content" style={FONT_STYLES.nanumMyeongjo}>
+                    코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요!
+                  </p>
+                  <span className="date">2023.07.08</span>
+                </a>
+              </li>
+              <li className="savedPostCard">
+                <a href="#">
+                  <div className="cardInfo">
+                    <div>
+                      <div className="photo"></div>
+                      <div className="fromName">
+                        <span>From. <em>김동훈</em></span>
+                        <Badge type="가족" />
+                      </div>
+                    </div>
+                    <a className="btnDelete">
+                      <img src={deleteIcon} alt="삭제하기" />
+                    </a>
+                  </div>
+                  <p className="content" style={FONT_STYLES.pretendard}>
+                    코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요!
+                  </p>
+                  <span className="date">2023.07.08</span>
+                </a>
+              </li>
+              <li className="savedPostCard">
+                <a href="#">
+                  <div className="cardInfo">
+                    <div>
+                      <div className="photo"></div>
+                      <div className="fromName">
+                        <span>From. <em>김동훈</em></span>
+                        <Badge type="가족" />
+                      </div>
+                    </div>
+                    <a className="btnDelete">
+                      <img src={deleteIcon} alt="삭제하기" />
+                    </a>
+                  </div>
+                  <p className="content" style={FONT_STYLES.NanumSonPyeonJiCe }>
+                    코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요!
+                  </p>
+                  <span className="date">2023.07.08</span>
+                </a>
+              </li>
+              <li className="savedPostCard">
+                <a href="#">
+                  <div className="cardInfo">
+                    <div>
+                      <div className="photo"></div>
+                      <div className="fromName">
+                        <span>From. <em>김동훈</em></span>
+                        <Badge type="동료" />
+                      </div>
+                    </div>
+                    <a className="btnDelete">
+                      <img src={deleteIcon} alt="삭제하기" />
+                    </a>
+                  </div>
+                  <p className="content" style={FONT_STYLES.NanumSonPyeonJiCe }>
+                    코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요!
+                  </p>
+                  <span className="date">2023.07.08</span>
+                </a>
+              </li>
+              <li className="savedPostCard">
+                <a href="#">
+                  <div className="cardInfo">
+                    <div>
+                      <div className="photo"></div>
+                      <div className="fromName">
+                        <span>From. <em>김동훈</em></span>
+                        <Badge type="지인" />
+                      </div>
+                    </div>
+                    <a className="btnDelete">
+                      <img src={deleteIcon} alt="삭제하기" />
+                    </a>
+                  </div>
+                  <p className="content" style={FONT_STYLES.NanumSonPyeonJiCe }>
+                    코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요!
+                  </p>
+                  <span className="date">2023.07.08</span>
+                </a>
+              </li>
+              <li className="savedPostCard">
+                <a href="#">
+                  <div className="cardInfo">
+                    <div>
+                      <div className="photo"></div>
+                      <div className="fromName">
+                        <span>From. <em>김동훈</em></span>
+                        <Badge type="가족" />
+                      </div>
+                    </div>
+                    <a className="btnDelete">
+                      <img src={deleteIcon} alt="삭제하기" />
+                    </a>
+                  </div>
+                  <p className="content" style={FONT_STYLES.NanumSonPyeonJiCe }>
+                    코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요!
+                  </p>
+                  <span className="date">2023.07.08</span>
+                </a>
+              </li>
+              <li className="savedPostCard">
+                <a href="#">
+                  <div className="cardInfo">
+                    <div>
+                      <div className="photo"></div>
+                      <div className="fromName">
+                        <span>From. <em>김동훈</em></span>
+                        <Badge type="가족" />
+                      </div>
+                    </div>
+                    <a className="btnDelete">
+                      <img src={deleteIcon} alt="삭제하기" />
+                    </a>
+                  </div>
+                  <p className="content" style={FONT_STYLES.NanumSonPyeonJiCe }>
+                    코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요!
+                  </p>
+                  <span className="date">2023.07.08</span>
+                </a>
+              </li>
+              <li className="savedPostCard">
+                <a href="#">
+                  <div className="cardInfo">
+                    <div>
+                      <div className="photo"></div>
+                      <div className="fromName">
+                        <span>From. <em>김동훈</em></span>
+                        <Badge type="가족" />
+                      </div>
+                    </div>
+                    <a className="btnDelete">
+                      <img src={deleteIcon} alt="삭제하기" />
+                    </a>
+                  </div>
+                  <p className="content" style={FONT_STYLES.NanumSonPyeonJiCe }>
+                    코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요. 건강, 체력 모두
+                    조심 또 하세요! 코로나가 또다시 기승을 부리는 요즘이네요.
+                    건강, 체력 모두 조심 또 하세요!
+                  </p>
+                  <span className="date">2023.07.08</span>
+                </a>
+              </li>
+            </ul>
+          </div>
+        </div>
+      </main>
+    </>
+  );
+}
+
+export default Post;
